Solution: Phase 1
Finish out _Phase 1_ of the solution, which involves a naive algorithm
to accomplish the on-read side of the problem.

Also, setup abstract tests that work against all `TaskAssigner`
implementations (or should!). This allows us to test both `TaskManager`
and `BlindTaskManager` for consistency, with a uniform set of tests that
take much less code to express.

In summary:
- Add abstract tests for classes adhering `TaskAssigner`
- Add tests to cover additional utility code
- Make the `TaskManager` constructor private, since using it versions
  the static factory method can be confusing